+++ codegen/smithy-aws-swift-codegen/src/main/kotlin/software/amazon/smithy/aws/swift/codegen/message/MessageMarshallableGenerator.kt
@@ -1,4 +1,4 @@
-package software.amazon.smithy.aws.swift.codegen
+package software.amazon.smithy.aws.swift.codegen.message
 
 import software.amazon.smithy.codegen.core.CodegenException
 import software.amazon.smithy.codegen.core.Symbol
